Room packets, part 2 episode 2
A deeper, albeit incomplete look into various packets that occur at the
end of shots, holes, and entire matches. A lot of strange things are
appearing, such as a packets whose formats change based on context
instead of a flag in the packet (GSS-0065, GSS-0066), and an appearance
of a 64-bit int in a 32-bit game (GSS-0066).

There are no depths too deep for this madness, ladies and gentlemen.