The Endgame
Various packets regarding ending of holes, matches and tournaments here;
GSC-0031 is similar to GSC-0006, but sent after every hole. GSC-00AA is
identical to GSC-0006, but sent when using a Tiki Report instead of
allowing a tournament to end naturally. The format for all three of
these packets have been split off into user_course_record_data.

While tournaments ending as a result of GSC-00AA is obvious, a new
method was needed for grouping packets that didn't directly result from
some other packet. Tentatively calling these 'events', the first such
event is that of a match or tournament ending, with GSC-0006 and 10 GSS
packets joining that list.

Lastly, GSS-012A is noted as a dedicated response to using a Tiki Report
with GSC-00AA, separate from it acting as a combination GSC-0006,
GSC-000F, and tournament end.