High Fashion
Details on user designed clothing in game, at least from an observer's
perspective. GSC-00B9 requests information on a particular item, and
GSS-012E allows for the client to make an HTTP request for custom
uploaded textures.

Also included is a small clean up of GSS-00E6, since user shops are the
most obvious places to find custom clothing on a whim.