More work on user info (#39)
* User Caddie Data

Following the User Character, Equipment, and Statistic Data structures,
User Caddie Data is similarly reused between multiple packets.

* Improve Change Equipment packets

More details for GSC-0006 and response GSS-006B.

* More User Statistics

Perhaps useless, but this is low hanging fruit.

User Statistics Data had to be split, as there is a 78-byte data block
of currently-unknown use at the end in GSS-0044 and GSS-0045 that does
not appear in GSS-0158.

* User Course Records Data

Another duplicated structure presents itself, this time with course
records. The extra course-specific slots of GSS-0045 appear in GSS-015C,
too.

* User Information III, The Reckoning

Continuing on the path of cleaning up GSC-002F responses, another
repeated data structure has been extracted (user_name_data), along with
pointing most of the GSC-002F responses to their appropriate shared
structure definitions instead of bespoke individual definitions that are
liable to be left incomplete when work happens elsewhere.

Because of the extensive use of these shared structures, a surprising
amount of extra information can be gathered about other accounts that
wouldn't be displayed in-game, such as current pang balance or exact
equipment loadout. Third party implementations of the server might see
fit to blank out this information where it's not absolutely necessary.

* Authentication Keep-Alive Observation

A minor observation that the current server's ID appears in the token
portion of the challenge and response. Also documented are the range of
lengths of responses.

* Statistical Anomaly

While GSS-0158 is mostly a response to GSC-002F, it's also sent as a
response to GSC-0002, but almost entirely blanked out.

Very odd.

* The Endgame

Various packets regarding ending of holes, matches and tournaments here;
GSC-0031 is similar to GSC-0006, but sent after every hole. GSC-00AA is
identical to GSC-0006, but sent when using a Tiki Report instead of
allowing a tournament to end naturally. The format for all three of
these packets have been split off into user_course_record_data.

While tournaments ending as a result of GSC-00AA is obvious, a new
method was needed for grouping packets that didn't directly result from
some other packet. Tentatively calling these 'events', the first such
event is that of a match or tournament ending, with GSC-0006 and 10 GSS
packets joining that list.

Lastly, GSS-012A is noted as a dedicated response to using a Tiki Report
with GSC-00AA, separate from it acting as a combination GSC-0006,
GSC-000F, and tournament end.

* Missed Connections

Additional packets related to connecting to a sub-server; GSC-0083 is
similar to GSC-0004, but used when the user is in multiplayer/grand prix
mode. With both GSC-0004 and GSC-0083, a reply of GSS-004E is used to
acknowledge a successful connection to the new sub-server.

* Minor fix to indices

Forgot to add GSC-0083 and GSS-004E to the packet indices.

* High Fashion

Details on user designed clothing in game, at least from an observer's
perspective. GSC-00B9 requests information on a particular item, and
GSS-012E allows for the client to make an HTTP request for custom
uploaded textures.

Also included is a small clean up of GSS-00E6, since user shops are the
most obvious places to find custom clothing on a whim.

* Port-a-Ball Code

GSC-0019 and GSS-0060 are the pair of packets responsible for allowing
players to relocate their comet during play, called relief. Note that
this is never free relief; it is always a penalty relief, regardless of
the impracticality/impossibility of playing the shot as the comet lies.

* Gotta Go Fast

GSC-0065 and GSS-00C7 are the pair of packets governing time booster /
fast forward use. GSC-0065 only occurs when the local user activates a
time booster, while GSS-00C7 is announced to everyone, thus actually
activating the time booster effect in the client.

* Masters of the Pangyaverse

GSC-0187, 0188, 0189, 018A, GSS-026E, 026F, 0270, and 0271 are all
related to Character Mastery.

GSC-018B, 018C, GSS-0272, and 0273 are said by other projects to also be
related to Character Mastery, but no examples of them have been
witnessed in order to be certain.

* Black Papel, hero of LaGacha

GSC-0186 and GSS-026C govern the 'big' mode for Black Papel, which are
otherwise identical to standard mode (GSC-014B and GSS-021B). A little
bit of cleanup on existing packets and an example of each type of result
is provided.