The great GSS-0076 (#46)
GSS-0076, type 0x04 is sent before VS matches to communicate to every
user in a room just about every common user data structure of all users
in said room, including themselves. This results in understandably large
packets (roughly 12KiB per player, with presumably up to 4 players).

It is likely that the clubset information is also a common structure,
but not yet seen in a previously documented packet. Once it's seen
again, it will be separated to its own file.

Included is one such example from a 3-player match. Note that
Packetdoc's web interface fails to properly parse/display this file,
likely due to its length.

Type 0x00 appears to be event related, and only contains a timestamp. It
is 23 bytes large. Also included is such an example.

Also took the opportunity to make the common data structures readily
accessible from the sidebar, and link all said structures to an index of
them instead of to each other.