My Room: A new explanation for why there are two user IDs. (#51)
I haven't confirmed this since I do not have a way to try earlier seasons, but I believe it is likely that the reason why there are two user IDs is due to the fact that in older seasons, it used to be possible to enter another user's room. Why the connecting user must specify their own user ID is still not totally known, but I think that most likely this solves the mystery of why there are two.

Even if this is true, they could possibly be in the other order. Hard to tell.